update permission control in API-rabbit

+++ api-rabbit/src/main/java/com/yunsoo/api/rabbit/controller/UserController.java
@@ -1,34 +1,29 @@
 package com.yunsoo.api.rabbit.controller;
 
+import com.yunsoo.api.rabbit.domain.UserDomain;
 import com.yunsoo.api.rabbit.dto.basic.User;
-import com.yunsoo.api.rabbit.object.TAccount;
-import com.yunsoo.api.rabbit.security.TokenAuthenticationService;
 import com.yunsoo.common.data.object.FileObject;
 import com.yunsoo.common.web.client.RestClient;
 import com.yunsoo.common.web.exception.BadRequestException;
 import com.yunsoo.common.web.exception.NotFoundException;
+import com.yunsoo.common.web.exception.UnauthorizedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.InputStreamResource;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.ByteArrayInputStream;
-import java.security.Principal;
 
 
 /**
  * Created by:   Zhe
  * Created on:   2015/3/3
  * Descriptions: This controller manage end user.
- *               Only authorized user can consume it.
+ * Only authorized user can consume it.
  * <p>
  * ErrorCode
  * 40401    :   User not found!
@@ -39,20 +34,21 @@ public class UserController {
 
     @Autowired
     private RestClient dataAPIClient;
-    private static final Logger LOGGER = LoggerFactory.getLogger(UserController.class);
     @Autowired
-    private TokenAuthenticationService tokenAuthenticationService;
+    private UserDomain userDomain;
+    private final String AUTH_HEADER_NAME = "YS_RABBIT_AUTH_TOKEN";
 
-//    @Autowired
-//    UserController(RestClient dataAPIClient) {
-//        this.dataAPIClient = dataAPIClient;
-//    }
+    private static final Logger LOGGER = LoggerFactory.getLogger(UserController.class);
 
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
-    @PreAuthorize("hasPermission(#user, 'user:read')")
-    public User getById(@PathVariable(value = "id") String id) throws NotFoundException {
+//    @PreAuthorize("hasPermission(#user, 'user:read')")
+    public User getById(@RequestHeader(AUTH_HEADER_NAME) String token,
+                        @PathVariable(value = "id") String id) throws NotFoundException {
         if (id == null || id.isEmpty()) {
             throw new BadRequestException("UserId不应为空！");
+        }
+        if (!userDomain.validateToken(token, id)) {
+            throw new UnauthorizedException("不能读取其他用户信息！");
         }
         User user = dataAPIClient.get("user/id/{id}", User.class, id);
         if (user == null) throw new NotFoundException(40401, "User not found id=" + id);
@@ -60,11 +56,15 @@ public class UserController {
     }
 
     @RequestMapping(value = "/cellular/{cellular}", method = RequestMethod.GET)
-    @PreAuthorize("hasPermission(#user, 'user:read')")
+//    @PreAuthorize("hasPermission(#user, 'user:read')")
 //    @PostFilter("hasPermission(filterObject, 'read') or hasPermission(filterObject, 'admin')")
-    public User getByCellular(@PathVariable(value = "cellular") String cellular) throws NotFoundException {
+    public User getByCellular(@RequestHeader(AUTH_HEADER_NAME) String token,
+                              @PathVariable(value = "cellular") String cellular) throws NotFoundException {
         if (cellular == null || cellular.isEmpty()) {
             throw new BadRequestException("cellular不能为空！");
+        }
+        if (!userDomain.validateToken(token)) {
+            throw new UnauthorizedException("不能读取其他用户信息！");
         }
         User user = dataAPIClient.get("user/cellular/{cellular}", User.class, cellular);
         if (user == null) throw new NotFoundException(40401, "User not found cellular=" + cellular);
@@ -73,8 +73,15 @@ public class UserController {
 
     @RequestMapping(value = "/device/{devicecode}", method = RequestMethod.GET)
 //    @PreAuthorize("hasAnyRole('COM_USER','YUNSOO_ADMIN')")
-    @PreAuthorize("hasPermission(#user, 'user:read')")
-    public User getByDevicecode(@PathVariable(value = "devicecode") String deviceCode) throws NotFoundException {
+//    @PreAuthorize("hasPermission(#user, 'user:read')")
+    public User getByDevicecode(@RequestHeader(AUTH_HEADER_NAME) String token,
+                                @PathVariable(value = "devicecode") String deviceCode) throws NotFoundException {
+        if (deviceCode == null || deviceCode.isEmpty()) {
+            throw new BadRequestException("deviceCode不能为空！");
+        }
+        if (!userDomain.validateToken(token)) {
+            throw new UnauthorizedException("不能读取其他用户信息！");
+        }
         User user = null;
         try {
             user = dataAPIClient.get("user/token/{devicecode}", User.class, deviceCode);
@@ -85,7 +92,7 @@ public class UserController {
     }
 
     @RequestMapping(value = "/thumbnail/{id}/{key}", method = RequestMethod.GET)
-    @PreAuthorize("hasPermission(#user, 'user:read')")
+//    @PreAuthorize("hasPermission(#user, 'user:read')")
     public ResponseEntity<?> getThumbnail(
             @PathVariable(value = "id") String id,
             @PathVariable(value = "key") String key) {
@@ -111,15 +118,23 @@ public class UserController {
 
 
     @RequestMapping(value = "", method = RequestMethod.PATCH)
-    @PreAuthorize("hasPermission(#user, 'user:update')")
-    public void updateUser(@RequestBody User user) throws Exception {
+//    @PreAuthorize("hasPermission(#user, 'user:update')")
+    public void updateUser(@RequestHeader(AUTH_HEADER_NAME) String token,
+                           @RequestBody User user) throws Exception {
+        if (!userDomain.validateToken(token, user.getId())) {
+            throw new UnauthorizedException("不能读取其他用户信息！");
+        }
         dataAPIClient.patch("user", user);
     }
 
     @RequestMapping(value = "/{userId}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PreAuthorize("hasPermission(#user, 'user:delete')")
-    public void deleteUser(@PathVariable(value = "userId") Long userId) throws Exception {
+//    @PreAuthorize("hasPermission(#user, 'user:delete')")
+    public void deleteUser(@RequestHeader(AUTH_HEADER_NAME) String token,
+                           @PathVariable(value = "userId") String userId) throws Exception {
+        if (!userDomain.validateToken(token, userId)) {
+            throw new UnauthorizedException("不能读取其他用户信息！");
+        }
         dataAPIClient.delete("user/{id}", userId);
     }
 }
